+++ src/worker/tests.rs
@@ -5,6 +5,7 @@ use crate::db::{self, test_support::ScopedTestDataDir};
 use super::{lock, mark_successful_job, record_failed_job_transition, recover_expired_jobs, run};
 use test_support::install_stub_codex;
 
+mod once_budget;
 pub(super) mod test_support;
 
 #[tokio::test]